requirements.txt
streamlit
google-generativeai
python-dotenv
altair==4.2.0